# Delta exporters will Export only those points which received new update