--- a/src/OpenTelemetry/CHANGELOG.md
+++ b/src/OpenTelemetry/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Metrics SDK will not provide inactive Metrics to exporter.
+* Metrics SDK will not provide inactive Metrics to delta exporter.
   ([#2629](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2629))
 
 * Histogram bounds are validated when added to a View.