--- a/src/OpenTelemetry/Metrics/BatchMetricPoint.cs
+++ b/src/OpenTelemetry/Metrics/BatchMetricPoint.cs
@@ -16,7 +16,6 @@
 
 using System;
 using System.Collections;
-using System.Diagnostics;
 using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Metrics
@@ -24,17 +23,18 @@ namespace OpenTelemetry.Metrics
     public readonly struct BatchMetricPoint : IDisposable
     {
         private readonly MetricPoint[] metricsPoints;
+        private readonly int[] metricPointsToProcess;
         private readonly long targetCount;
         private readonly DateTimeOffset start;
         private readonly DateTimeOffset end;
 
-        internal BatchMetricPoint(MetricPoint[] metricsPoints, int maxSize, DateTimeOffset start, DateTimeOffset end)
+        internal BatchMetricPoint(MetricPoint[] metricsPoints, int[] metricPointsToProcess, long targetCount, DateTimeOffset start, DateTimeOffset end)
         {
-            Debug.Assert(maxSize > 0, $"{nameof(maxSize)} should be a positive number.");
             Guard.Null(metricsPoints, nameof(metricsPoints));
 
             this.metricsPoints = metricsPoints;
-            this.targetCount = maxSize;
+            this.metricPointsToProcess = metricPointsToProcess;
+            this.targetCount = targetCount;
             this.start = start;
             this.end = end;
         }
@@ -50,7 +50,7 @@ namespace OpenTelemetry.Metrics
         /// <returns><see cref="Enumerator"/>.</returns>
         public Enumerator GetEnumerator()
         {
-            return new Enumerator(this.metricsPoints, this.targetCount, this.start, this.end);
+            return new Enumerator(this.metricsPoints, this.metricPointsToProcess, this.targetCount, this.start, this.end);
         }
 
         /// <summary>
@@ -59,14 +59,16 @@ namespace OpenTelemetry.Metrics
         public struct Enumerator : IEnumerator
         {
             private readonly MetricPoint[] metricsPoints;
+            private readonly int[] metricPointsToProcess;
             private readonly DateTimeOffset start;
             private readonly DateTimeOffset end;
             private long targetCount;
             private long index;
 
-            internal Enumerator(MetricPoint[] metricsPoints, long targetCount, DateTimeOffset start, DateTimeOffset end)
+            internal Enumerator(MetricPoint[] metricsPoints, int[] metricPointsToProcess, long targetCount, DateTimeOffset start, DateTimeOffset end)
             {
                 this.metricsPoints = metricsPoints;
+                this.metricPointsToProcess = metricPointsToProcess;
                 this.targetCount = targetCount;
                 this.index = -1;
                 this.start = start;
@@ -77,7 +79,7 @@ namespace OpenTelemetry.Metrics
             {
                 get
                 {
-                    return ref this.metricsPoints[this.index];
+                    return ref this.metricsPoints[this.metricPointsToProcess[this.index]];
                 }
             }
 
@@ -93,12 +95,7 @@ namespace OpenTelemetry.Metrics
             {
                 while (++this.index < this.targetCount)
                 {
-                    ref var metricPoint = ref this.metricsPoints[this.index];
-                    if (metricPoint.StartTime == default)
-                    {
-                        continue;
-                    }
-
+                    ref var metricPoint = ref this.metricsPoints[this.metricPointsToProcess[this.index]];
                     metricPoint.StartTime = this.start;
                     metricPoint.EndTime = this.end;
                     return true;