@cijothomas 2021-11-17
This is almost same as what @alanwest did the Metric Point clean up PR, except that this PR does not do the actual clean up/reuse.